requirements.txt
easydev>=0.9.34
pandas
bioservices>=1.4.5
colormap
scipy
biopython
matplotlib
numpydoc
colorlog
